+++ controller/rq_selectionRange.mli
@@ -15,4 +15,4 @@
 (* Written by: Emilio J. Gallego Arias                                  *)
 (************************************************************************)
 
-val request : Request.position
+val request : points:(int * int) list -> Request.position